Cut corners when processing watch events
Construct fake watch persister entries and
add them to the persister. This is to avoid
going through huge lists of watchers (and do
thousands of API calls) just to retrieve a single 
watcher when processing an event.

GitHub also restricts watcher lists to 400 pages
(with per_page=100) for big projects.