Sort files in MySQL dump archive by size
When the tarball is streamed and unpacked on the fly, it makes sense to extract the smallest files first. 
Apart from that, the directory name is preserved and the overall behavior does not change.
`$file` should not be put into quotes as `tar` will interpret the whole list of files as a single file and fail.